api/pyproject.toml
[project]
name = "schoolmarks-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery==5.4.0",
    "dj-rest-auth==7.0.1",
    "django-allauth==65.4.1",
    "django-cors-headers==4.7.0",
    "django-rest-framework==0.1.0",
    "djangorestframework-simplejwt==5.4.0",
    "drf-spectacular==0.28.0",
    "drf-yasg==1.21.8",
    "fido2==1.2.0",
    "gunicorn==23.0.0",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "redis==5.2.1",
    "requests==2.32.3",
    "setuptools==75.8.0",
    "typing-extensions==4.12.2",
]

[dependency-groups]
dev = [
    "ruff>=0.9.7",
]
